rename: repository, exception 폴더 생성 및 정리 #3
