+++ src/main/java/com/smartlogis/userservice/domain/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.smartlogis.userservice.domain;
+package com.smartlogis.userservice.domain.repository;
 
 import java.util.Optional;
 
@@ -6,6 +6,8 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.repository.Repository;
 
+import com.smartlogis.userservice.domain.User;
+import com.smartlogis.userservice.domain.UserId;
 import com.smartlogis.userservice.domain.dto.UserSearch;
 
 public interface UserRepository extends Repository<User, UserId>, UserRepositoryCustom {
