+++ src/main/java/com/smartlogis/userservice/domain/repository/UserRepositoryCustom.java
@@ -1,8 +1,9 @@
-package com.smartlogis.userservice.domain;
+package com.smartlogis.userservice.domain.repository;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import com.smartlogis.userservice.domain.User;
 import com.smartlogis.userservice.domain.dto.UserSearch;
 
 public interface UserRepositoryCustom {
