+++ src/main/java/com/smartlogis/userservice/domain/service/UserServiceImpl.java
@@ -5,10 +5,10 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.validation.annotation.Validated;
 
 import com.smartlogis.userservice.domain.User;
-import com.smartlogis.userservice.domain.UserException;
+import com.smartlogis.userservice.domain.exception.UserException;
 import com.smartlogis.userservice.domain.UserId;
-import com.smartlogis.userservice.domain.UserMessageCode;
-import com.smartlogis.userservice.domain.UserRepository;
+import com.smartlogis.userservice.domain.exception.UserMessageCode;
+import com.smartlogis.userservice.domain.repository.UserRepository;
 import com.smartlogis.userservice.domain.UserRole;
 import com.smartlogis.userservice.domain.dto.OrganizationInfo;
 import com.smartlogis.userservice.domain.dto.UserCreate;
